// apps/chat-ui/src/server/openapi-client/models/ContinueChatRequestDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Chat API
 * Multi-turn conversation API built with LangChain and LangGraph. This API maintains conversation history using LangGraph\'s MemorySaver, allowing for contextual multi-turn conversations. The assistant has access to user information tools for personalization.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ContinueChatRequestDto
 */
export interface ContinueChatRequestDto {
    /**
     * The message from the user to continue the conversation
     * @type {string}
     * @memberof ContinueChatRequestDto
     */
    message: string;
    /**
     * The name of the user sending the message. Used for personalization.
     * @type {string}
     * @memberof ContinueChatRequestDto
     */
    user: string;
}

/**
 * Check if a given object implements the ContinueChatRequestDto interface.
 */
export function instanceOfContinueChatRequestDto(value: object): value is ContinueChatRequestDto {
    if (!('message' in value) || value['message'] === undefined) return false;
    if (!('user' in value) || value['user'] === undefined) return false;
    return true;
}

export function ContinueChatRequestDtoFromJSON(json: any): ContinueChatRequestDto {
    return ContinueChatRequestDtoFromJSONTyped(json, false);
}

export function ContinueChatRequestDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ContinueChatRequestDto {
    if (json == null) {
        return json;
    }
    return {
        
        'message': json['message'],
        'user': json['user'],
    };
}

export function ContinueChatRequestDtoToJSON(json: any): ContinueChatRequestDto {
    return ContinueChatRequestDtoToJSONTyped(json, false);
}

export function ContinueChatRequestDtoToJSONTyped(value?: ContinueChatRequestDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'message': value['message'],
        'user': value['user'],
    };
}